+++ packages/contracts/contracts/registry/DAORegistry.sol
@@ -15,7 +15,7 @@ contract DAORegistry is InterfaceBasedRegistry {
     bytes32 public constant REGISTER_DAO_PERMISSION_ID = keccak256("REGISTER_DAO_PERMISSION");
 
     /// @notice The ENS subdomain registrar registering the DAO subdomains.
-    ENSSubdomainRegistrar private subdomainRegistrar;
+    ENSSubdomainRegistrar public subdomainRegistrar;
 
     /// @notice Thrown if the DAO subdomain doesn't match the regex `[0-9a-z\-]`
     error InvalidDaoSubdomain(string subdomain);