+++ packages/contracts/test/voting/addresslist.ts
@@ -3,7 +3,7 @@ import {ethers} from 'hardhat';
 import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 
 import {AddresslistMock} from '../../typechain';
-import {ADDRESSLIST_EVENTS} from '../../utils/event';
+import {MEMBERSHIP_EVENTS} from '../../utils/event';
 
 describe('AddresslistMock', function () {
   let signers: SignerWithAddress[];
@@ -47,6 +47,7 @@ describe('AddresslistMock', function () {
       expect(await addresslist.addresslistLength()).to.equal(0);
     });
   });
+
   context('addresslistLengthAtBlock', function () {
     it('returns the right length after addresses were added', async () => {
       let tx1 = await addresslist.addAddresses([signers[0].address]);
@@ -162,10 +163,10 @@ describe('AddresslistMock', function () {
       expect(await addresslist.addresslistLength()).to.equal(2);
     });
 
-    it('emits the `AddressesAdded` event', async () => {
+    it('emits the `MembersAdded` event', async () => {
       let addresses = [signers[0].address, signers[1].address];
       await expect(addresslist.addAddresses(addresses))
-        .to.emit(addresslist, ADDRESSLIST_EVENTS.ADDRESSES_ADDED)
+        .to.emit(addresslist, MEMBERSHIP_EVENTS.MEMBERS_ADDED)
         .withArgs(addresses);
     });
 
@@ -209,7 +210,7 @@ describe('AddresslistMock', function () {
       await expect(
         addresslist.removeAddresses([signers[0].address, signers[1].address])
       )
-        .to.emit(addresslist, ADDRESSLIST_EVENTS.ADDRESSES_REMOVED)
+        .to.emit(addresslist, MEMBERSHIP_EVENTS.MEMBERS_REMOVED)
         .withArgs([signers[0].address, signers[1].address]);
 
       await ethers.provider.send('evm_mine', []);
@@ -219,14 +220,14 @@ describe('AddresslistMock', function () {
       expect(await addresslist.addresslistLength()).to.equal(0);
     });
 
-    it('emits the `AddressesRemoved` event', async () => {
+    it('emits the `MembersRemoved` event', async () => {
       let addresses = [signers[0].address, signers[1].address];
 
       await addresslist.addAddresses(addresses);
       await ethers.provider.send('evm_mine', []);
 
       await expect(addresslist.removeAddresses(addresses))
-        .to.emit(addresslist, ADDRESSLIST_EVENTS.ADDRESSES_REMOVED)
+        .to.emit(addresslist, MEMBERSHIP_EVENTS.MEMBERS_REMOVED)
         .withArgs(addresses);
     });