+++ packages/subgraph/src/packages/multisig/multisig.ts
@@ -3,8 +3,8 @@ import {dataSource, store, BigInt} from '@graphprotocol/graph-ts';
 import {
   ProposalCreated,
   ProposalExecuted,
-  AddressesAdded,
-  AddressesRemoved,
+  MembersAdded,
+  MembersRemoved,
   Multisig,
   Approved,
   MultisigSettingsUpdated
@@ -158,7 +158,7 @@ export function handleProposalExecuted(event: ProposalExecuted): void {
   }
 }
 
-export function handleAddressesAdded(event: AddressesAdded): void {
+export function handleMembersAdded(event: MembersAdded): void {
   const members = event.params.members;
   for (let index = 0; index < members.length; index++) {
     const member = members[index].toHexString();
@@ -175,7 +175,7 @@ export function handleAddressesAdded(event: AddressesAdded): void {
   }
 }
 
-export function handleAddressesRemoved(event: AddressesRemoved): void {
+export function handleMembersRemoved(event: MembersRemoved): void {
   const members = event.params.members;
   for (let index = 0; index < members.length; index++) {
     const member = members[index].toHexString();