+++ packages/subgraph/src/packages/token/token-voting.ts
@@ -5,6 +5,7 @@ import {
   ProposalCreated,
   ProposalExecuted,
   VotingSettingsUpdated,
+  MembershipContractAnnounced,
   TokenVoting
 } from '../../../generated/templates/TokenVoting/TokenVoting';
 import {
@@ -14,6 +15,7 @@ import {
   TokenVotingVoter,
   TokenVotingVote
 } from '../../../generated/schema';
+import {handleERC20Token} from '../../utils/tokens';
 
 import {RATIO_BASE, VOTER_OPTIONS, VOTING_MODES} from '../../utils/constants';
 
@@ -239,6 +241,19 @@ export function handleVotingSettingsUpdated(
     packageEntity.minParticipation = event.params.minParticipation;
     packageEntity.minDuration = event.params.minDuration;
     packageEntity.minProposerVotingPower = event.params.minProposerVotingPower;
+    packageEntity.save();
+  }
+}
+
+export function handleMembershipContractAnnounced(
+  event: MembershipContractAnnounced
+): void {
+  let token = event.params.definingContract;
+  let packageEntity = TokenVotingPlugin.load(event.address.toHexString());
+
+  if (packageEntity) {
+    packageEntity.token = handleERC20Token(token);
+
     packageEntity.save();
   }
 }