+++ packages/subgraph/tests/addresslist-voting/addresslist-voting.test.ts
@@ -2,10 +2,10 @@ import {assert, clearStore, test} from 'matchstick-as/assembly/index';
 import {Address, BigInt} from '@graphprotocol/graph-ts';
 
 import {
-  handleAddressesAdded,
+  handleMembersAdded,
   handleVoteCast,
   handleProposalExecuted,
-  handleAddressesRemoved,
+  handleMembersRemoved,
   handleVotingSettingsUpdated,
   _handleProposalCreated
 } from '../../src/packages/addresslist/addresslist-voting';
@@ -40,10 +40,10 @@ import {
   createTotalVotingPowerCall
 } from '../utils';
 import {
-  createNewAddressesAddedEvent,
+  createNewMembersAddedEvent,
   createNewVoteCastEvent,
   createNewProposalExecutedEvent,
-  createNewAddressesRemovedEvent,
+  createNewMembersRemovedEvent,
   createNewProposalCreatedEvent,
   createNewVotingSettingsUpdatedEvent,
   getProposalCountCall,
@@ -548,17 +548,17 @@ test('Run AddresslistVoting (handleVotingSettingsUpdated) mappings with mock eve
   clearStore();
 });
 
-test('Run AddresslistVoting (handleAddressesAdded) mappings with mock event', () => {
+test('Run AddresslistVoting (handleMembersAdded) mappings with mock event', () => {
   let userArray = [
     Address.fromString(ADDRESS_ONE),
     Address.fromString(ADDRESS_TWO)
   ];
 
   // create event
-  let event = createNewAddressesAddedEvent(userArray, CONTRACT_ADDRESS);
+  let event = createNewMembersAddedEvent(userArray, CONTRACT_ADDRESS);
 
   // handle event
-  handleAddressesAdded(event);
+  handleMembersAdded(event);
 
   // checks
 
@@ -584,7 +584,7 @@ test('Run AddresslistVoting (handleAddressesAdded) mappings with mock event', ()
   clearStore();
 });
 
-test('Run AddresslistVoting (AddressesRemoved) mappings with mock event', () => {
+test('Run AddresslistVoting (MembersRemoved) mappings with mock event', () => {
   // create state
   let memberAddresses = [
     Address.fromString(ADDRESS_ONE),
@@ -614,13 +614,13 @@ test('Run AddresslistVoting (AddressesRemoved) mappings with mock event', () => 
   assert.fieldEquals('AddresslistVotingVoter', memberId2, 'id', memberId2);
 
   // create event
-  let event = createNewAddressesRemovedEvent(
+  let event = createNewMembersRemovedEvent(
     [memberAddresses[1]],
     CONTRACT_ADDRESS
   );
 
   // handle event
-  handleAddressesRemoved(event);
+  handleMembersRemoved(event);
 
   // checks
   assert.fieldEquals('AddresslistVotingVoter', memberId1, 'id', memberId1);